// src/modules/admin/tournament/utils/type.ts
export enum StandardColor {
  Black = 'black',
  White = 'white',
}

export enum ClothesSize {
  S = 's',
  M = 'm',
  L = 'l',
  XL = 'xl',
  XL_2 = '2xl',
  XL_3 = '3xl',
  XL_4 = '4xl',
}

export enum Country {
  Vietnam = 'Vietnam',
  Singapore = 'Singapore',
  Cambodia = 'Cambodia',
  Malaysia = 'Malaysia',
  Philipines = 'Philipines',
  Other = 'Other',
}